+++ assets/js/dashboard/nav-menu/top-bar.tsx
@@ -49,12 +49,12 @@ function TopBarStickyWrapper({ children }: { children: ReactNode }) {
 function TopBarInner({ showCurrentVisitors }: TopBarProps) {
   const site = useSiteContext()
   const user = useUserContext()
-  const { saved_segments } = site.flags
+  const { saved_segments, saved_segments_fe } = site.flags
   const leftActionsRef = useRef<HTMLDivElement>(null)
 
   return (
     <div className="flex items-center w-full">
-      {saved_segments ? (
+      {!!saved_segments && !!saved_segments_fe ? (
         <>
           <div
             className="flex items-center gap-x-4 shrink-0"