+++ lib/plausible_web/plugs/authorize_site_access.ex
@@ -40,8 +40,6 @@ defmodule PlausibleWeb.Plugs.AuthorizeSiteAccess do
 
   @all_roles [:public, :viewer, :admin, :editor, :super_admin, :owner]
 
-  def all_roles(), do: @all_roles
-
   def init([]), do: {@all_roles, nil}
 
   def init(:all_roles), do: {@all_roles, nil}